requirements.txt
# UI Automation and Interaction
pyautogui
# Computer Vision and OCR
easyocr
opencv-python
# Image Processing
Pillow
# HTTP client
requests
# Ollama
ollama
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langchain>=0.1.0

